#6 Use Case: Generate Report Further implementation of use case. In fact, working solution for generating excel reports.

+++ domain/src/main/java/de/cassisi/hearth/entity/PerfusorData.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public final class PerfusorData {
 
     private final String name;
-    private final double rate;
+    private final int rate;
 
 }
